+++ crates/aranya-client-capi/src/imp/config.rs
@@ -4,42 +4,13 @@ use aranya_capi_core::{
     safe::{TypeId, Typed},
     InvalidArg,
 };
-use buggy::bug;
 
 use crate::api::defs::Duration;
 
 #[repr(C)]
-#[derive(Debug, Copy, Clone)]
-/// Configuration values for syncing with a peer
-pub struct SyncPeerConfig {
-    interval: Duration,
-    sync_now: bool,
-}
-
-impl Typed for SyncPeerConfig {
-    const TYPE_ID: TypeId = TypeId::new(0x2049e682);
-}
-
-impl From<SyncPeerConfig> for aranya_client::client::SyncPeerConfig {
-    fn from(value: SyncPeerConfig) -> Self {
-        Self::builder()
-            .interval(value.interval.into())
-            .sync_now(value.sync_now)
-            .build()
-            .expect("All values are set")
-    }
-}
-
-impl From<&SyncPeerConfig> for aranya_client::client::SyncPeerConfig {
-    fn from(value: &SyncPeerConfig) -> Self {
-        (*value).into()
-    }
-}
-
-#[repr(C)]
 #[derive(Copy, Clone, Debug)]
 #[aranya_capi_core::opaque(size = 32, align = 8)]
-/// Configuration info for Aranya
+/// Configuration info when creating a client in Aranya
 pub struct ClientConfig {
     pub daemon_addr: *const c_char,
     #[cfg(feature = "afc")]
@@ -61,17 +32,25 @@ pub struct ClientConfigBuilder {
 }
 
 impl ClientConfigBuilder {
-    /// Attempts to construct a [`ClientConfig`], returning an [`Error::Bug`](super::Error::Bug) if
-    /// there are invalid parameters.
+    /// Attempts to construct a [`ClientConfig`], returning an
+    /// [`Error::InvalidArg`](super::Error::InvalidArg) if there are invalid parameters.
     pub fn build(self) -> Result<ClientConfig, super::Error> {
         if self.daemon_addr.is_null() {
-            bug!("Tried to create a ClientConfig without a valid address!");
+            let e = InvalidArg::new(
+                "daemon_addr",
+                "Tried to create a `ClientConfig` without setting the daemon address!",
+            );
+            return Err(e.into());
         }
 
         #[cfg(feature = "afc")]
         let Some(afc) = self.afc
         else {
-            bug!("Tried to create a ClientConfig without a valid AfcConfig!");
+            let e = InvalidArg::new(
+                "afc_config",
+                "Tried to create a `ClientConfig` without setting a valid `AfcConfig`!",
+            );
+            return Err(e.into());
         };
 
         Ok(ClientConfig {
@@ -79,57 +58,6 @@ impl ClientConfigBuilder {
             #[cfg(feature = "afc")]
             afc,
         })
-    }
-}
-
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-/// Builder for a [`SyncPeerConfig`]
-pub struct SyncPeerConfigBuilder {
-    interval: Option<Duration>,
-    sync_now: bool,
-}
-
-impl SyncPeerConfigBuilder {
-    /// Set the interval at which syncing occurs
-    pub fn interval(&mut self, duration: Duration) {
-        self.interval = Some(duration);
-    }
-
-    /// Configures whether the peer will be immediately synced with after being added.
-    ///
-    /// By default, the peer is immediately synced with.
-    pub fn sync_now(&mut self, sync_now: bool) {
-        self.sync_now = sync_now;
-    }
-
-    /// Build a [`SyncPeerConfig`]
-    pub fn build(&self) -> Result<SyncPeerConfig, super::Error> {
-        let Some(interval) = self.interval else {
-            let e = Into::into(InvalidArg::new(
-                "interval",
-                "Tried to create a `SyncPeerConfig` without setting the interval!",
-            ));
-            return Err(e);
-        };
-
-        Ok(SyncPeerConfig {
-            interval,
-            sync_now: self.sync_now,
-        })
-    }
-}
-
-impl Typed for SyncPeerConfigBuilder {
-    const TYPE_ID: TypeId = TypeId::new(0x2049e683);
-}
-
-impl Default for SyncPeerConfigBuilder {
-    fn default() -> Self {
-        Self {
-            interval: None,
-            sync_now: true,
-        }
     }
 }
 
@@ -167,15 +95,23 @@ pub struct AfcConfigBuilder {
 
 #[cfg(feature = "afc")]
 impl AfcConfigBuilder {
-    /// Attempts to construct an [`AfcConfig`], returning an [`Error::Bug`](super::Error::Bug) if
-    /// there are invalid parameters.
+    /// Attempts to construct an [`AfcConfig`], returning an
+    /// [`Error::InvalidArg`](super::Error::InvalidArg) if there are invalid parameters.
     pub fn build(self) -> Result<AfcConfig, super::Error> {
         if self.shm_path.is_null() {
-            bug!("Tried to create an AfcConfig without a valid shm_path!");
+            let e = InvalidArg::new(
+                "shm_path",
+                "Tried to create a `AfcConfig` without setting a valid shared memory path!",
+            );
+            return Err(e.into());
         }
 
         if self.addr.is_null() {
-            bug!("Tried to create an AfcConfig without a valid address!");
+            let e = InvalidArg::new(
+                "address",
+                "Tried to create a `AfcConfig` without setting a valid address!",
+            );
+            return Err(e.into());
         }
 
         Ok(AfcConfig {
@@ -188,8 +124,88 @@ impl AfcConfigBuilder {
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
+/// Configuration info for syncing with a peer
+pub struct SyncPeerConfig {
+    interval: Duration,
+    sync_now: bool,
+}
+
+impl Typed for SyncPeerConfig {
+    const TYPE_ID: TypeId = TypeId::new(0x2049E682);
+}
+
+impl From<SyncPeerConfig> for aranya_client::SyncPeerConfig {
+    fn from(value: SyncPeerConfig) -> Self {
+        Self::builder()
+            .interval(value.interval.into())
+            .sync_now(value.sync_now)
+            .build()
+            .expect("All values are set")
+    }
+}
+
+impl From<&SyncPeerConfig> for aranya_client::SyncPeerConfig {
+    fn from(value: &SyncPeerConfig) -> Self {
+        (*value).into()
+    }
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug)]
+/// Builder for a [`SyncPeerConfig`]
+pub struct SyncPeerConfigBuilder {
+    interval: Option<Duration>,
+    sync_now: bool,
+}
+
+impl SyncPeerConfigBuilder {
+    /// Sets the interval at which syncing occurs.
+    pub fn interval(&mut self, duration: Duration) {
+        self.interval = Some(duration);
+    }
+
+    /// Configures whether the peer will be immediately synced with after being added.
+    ///
+    /// By default, the peer is immediately synced with.
+    pub fn sync_now(&mut self, sync_now: bool) {
+        self.sync_now = sync_now;
+    }
+
+    /// Attempts to construct a [`SyncPeerConfig`], returning an
+    /// [`Error::InvalidArg`](super::Error::InvalidArg) if there are invalid parameters.
+    pub fn build(&self) -> Result<SyncPeerConfig, super::Error> {
+        let Some(interval) = self.interval else {
+            let e = InvalidArg::new(
+                "interval",
+                "Tried to create a `SyncPeerConfig` without setting the interval!",
+            );
+            return Err(e.into());
+        };
+
+        Ok(SyncPeerConfig {
+            interval,
+            sync_now: self.sync_now,
+        })
+    }
+}
+
+impl Typed for SyncPeerConfigBuilder {
+    const TYPE_ID: TypeId = TypeId::new(0x2049E683);
+}
+
+impl Default for SyncPeerConfigBuilder {
+    fn default() -> Self {
+        Self {
+            interval: None,
+            sync_now: true,
+        }
+    }
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug)]
 #[aranya_capi_core::opaque(size = 0, align = 1)]
-/// Configuration info for Aranya
+/// Configuration info when creating or adding a team in Aranya
 pub struct TeamConfig {}
 
 impl Typed for TeamConfig {
