+++ crates/aranya-client/src/client.rs
@@ -1,6 +1,6 @@
 //! Client-daemon connection.
 
-use std::{net::SocketAddr, path::Path, time::Duration};
+use std::{net::SocketAddr, path::Path};
 
 #[cfg(feature = "afc")]
 use aranya_daemon_api::CS;
@@ -20,7 +20,7 @@ use tracing::{debug, info, instrument};
 #[cfg(feature = "afc")]
 use crate::afc::{setup_afc_shm, FastChannels, FastChannelsImpl};
 use crate::{
-    config::TeamConfig,
+    config::{SyncPeerConfig, TeamConfig},
     error::{Error, Result},
 };
 
@@ -158,18 +158,18 @@ impl Client {
 
     /// Create a new graph/team with the current device as the owner.
     pub async fn create_team(&mut self, cfg: TeamConfig) -> Result<TeamId> {
-        Ok(self
-            .daemon
+        self.daemon
             .create_team(context::current(), cfg.into())
-            .await??)
+            .await?
+            .map_err(Into::into)
     }
 
     /// Add a team to the local device store.
     pub async fn add_team(&mut self, team: TeamId, cfg: TeamConfig) -> Result<()> {
-        Ok(self
-            .daemon
+        self.daemon
             .add_team(context::current(), team, cfg.into())
-            .await??)
+            .await?
+            .map_err(Into::into)
     }
 
     /// Remove a team from the local device store.
@@ -459,84 +459,5 @@ impl Queries<'_> {
             .query_label_exists(context::current(), self.id, label)
             .await?
             .map_err(Into::into)
-    }
-}
-
-/// Configuration values for syncing with a peer
-#[derive(Debug, Clone)]
-pub struct SyncPeerConfig {
-    interval: Duration,
-    sync_now: bool,
-}
-
-impl SyncPeerConfig {
-    /// Creates a default [`SyncPeerConfigBuilder`]
-    pub fn builder() -> SyncPeerConfigBuilder {
-        Default::default()
-    }
-}
-
-impl From<SyncPeerConfig> for aranya_daemon_api::SyncPeerConfig {
-    fn from(value: SyncPeerConfig) -> Self {
-        Self {
-            interval: value.interval,
-            sync_now: value.sync_now,
-        }
-    }
-}
-
-/// Builder for a [`SyncPeerConfig`]
-pub struct SyncPeerConfigBuilder {
-    interval: Option<Duration>,
-    sync_now: bool,
-}
-
-impl SyncPeerConfigBuilder {
-    /// Creates a `SyncPeerConfigBuilder`.
-    pub fn new() -> Self {
-        Default::default()
-    }
-
-    /// Build a [`SyncPeerConfig`]
-    pub fn build(self) -> Result<SyncPeerConfig> {
-        let Some(interval) = self.interval else {
-            let e = Error::InvalidArg {
-                arg: "interval",
-                reason: "Tried to create a `SyncPeerConfig` without setting the interval!",
-            };
-            return Err(e);
-        };
-
-        Ok(SyncPeerConfig {
-            interval,
-            sync_now: self.sync_now,
-        })
-    }
-
-    /// Set the interval at which syncing occurs
-    ///
-    /// By default, the interval is not set. It is an error to call
-    /// [`build`][Self::build] before setting the interval with
-    /// this method
-    pub fn interval(mut self, duration: Duration) -> Self {
-        self.interval = Some(duration);
-        self
-    }
-
-    /// Configures whether the peer will be immediately synced with after being added.
-    ///
-    /// By default, the peer is immediately synced with.
-    pub fn sync_now(mut self, sync_now: bool) -> Self {
-        self.sync_now = sync_now;
-        self
-    }
-}
-
-impl Default for SyncPeerConfigBuilder {
-    fn default() -> Self {
-        Self {
-            interval: None,
-            sync_now: true,
-        }
     }
 }
